+++ test/widgets/multi_day_drag_over_body_test.dart
@@ -1,4 +1,3 @@
-import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:kalender/kalender.dart';
 import 'package:kalender/src/widgets/event_tiles/tiles/multi_day_tile.dart' show MultiDayEventTile;
@@ -43,7 +42,7 @@ void main() {
         callbacks: callbacks,
         viewConfiguration: MultiDayViewConfiguration.week(
           displayRange: year2025DisplayRange,
-          initialTimeOfDay: const TimeOfDay(hour: 0, minute: 0),
+          initialTimeOfDay: const KalenderTime(hour: 0, minute: 0),
           initialDateTime: start,
         ),
         header: CalendarHeader(interaction: interaction),